moved execute_query to external
